You can now access the PCSX menu by pressing [select] + [triangle] on your control pad.
Folder structure was changed to keep the hack and the backup separated.

It is possible to overmount the original games folder and show only the games on the USB drive.

A new file (lolhack_template.sh) was added to be the base of a GUI to make the process of backup/restore the system and adding/removing games easier.